refactor du frontend Discord : utilisation du context
